refactor: extract save actor and command to save directory
