+++ src/services/statefuls/snapshot/mod.rs
@@ -1,7 +1,6 @@
 pub mod dump_loader;
-pub mod save_command;
 pub mod endec;
-pub(crate) mod save_actor;
+pub mod save;
 
 use super::cache::CacheEntry;
